refactor: Moved children components of the exportable into a single file for improved organization, renamed files, and extracted scrolling buttons into a globally resuable component.
